requirements.txt
# Core ML libraries
torch==2.4.1
torchvision==0.19.1
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
opencv-python==4.10.0.84
tqdm==4.66.4
pillow==10.3.0

# Optional: Development tools
# jupyter==1.0.0
# pytest==7.0.0
